removed bug - adding move history
the bug was because of an obsolete forEach in piece.update,
which I now deleted

instead I am adding a history of moves.
this will solve a lot of problems in adding en passant mechanics.

because of that, I have paused my coding of the en passant mechanics.
first I will write a workable history.
then I will use that to further "en passant".